+++ lib/internal/quic/core.js
@@ -262,8 +262,7 @@ function onSessionClose(code, family) {
 // some basic information about the ALPN, SNI, and Ciphers that are
 // being requested. It is only called if the 'clientHello' event is
 // listened for.
-function onSessionClientHello(alpn, servername, ciphers, callback) {
-  callback = callback.bind(this);
+function onSessionClientHello(alpn, servername, ciphers) {
   this[owner_symbol][kClientHello](
     alpn,
     servername,
@@ -274,7 +273,7 @@ function onSessionClientHello(alpn, servername, ciphers, callback) {
         return;
       }
       try {
-        callback(...args);
+        this.onClientHelloDone(...args);
       } catch (err) {
         this[owner_symbol].destroy(err);
       }
@@ -283,10 +282,9 @@ function onSessionClientHello(alpn, servername, ciphers, callback) {
 
 // This callback is only ever invoked for QuicServerSession instances,
 // and is used to trigger OCSP request processing when needed. The
-// user callback must invoke the callback function in order for the
+// user callback must invoke .onCertDone() in order for the
 // TLS handshake to continue.
-function onSessionCert(servername, callback) {
-  callback = callback.bind(this);
+function onSessionCert(servername) {
   this[owner_symbol][kCert](servername, (err, context, ocspResponse) => {
     if (err) {
       this[owner_symbol].destroy(err);
@@ -311,7 +309,7 @@ function onSessionCert(servername, callback) {
       }
     }
     try {
-      callback(context ? context.context : undefined, ocspResponse);
+      this.onCertDone(context ? context.context : undefined, ocspResponse);
     } catch (err) {
       this[owner_symbol].destroy(err);
     }
